refactor: move the headers parser out to the command package
Just moving things around to keep the main package as lean as possible.
